chore!: change fd_event_callback signature to take an events flag
It was assumed before this the events returned were always the events
that were asked via evmask. But of course, this is not necessarily true,
since epoll could return an error condition on a given socket, for
example, and then the callbacks have no way to check for this case.
The current change fixes this shortcoming.

Note there is a currently a single flag - FD_EVENT_ERROR - that can be
checked to see if there is an error condition. More (granular) flags
could be added later if needed.

The commit also changes existing code to respect the updated
fd_Event_callback signature.

Some changes are also made in examples etc to silence compiler warnings
(treated as errors)