chore: fix CMakeLists.txt installation
- before this commit the cmakelists would automatically append usr/ to
what it always assumed was a root path; however, this leads to
problems if the user specifies the usr -- which the yocto build system
does, for example. As a result you end up with e.g. usr/usr/bin and
other nonsensical paths.
With this commit, the user *must* specify 'usr' if that's where libtarp
is to be installed.

E.g.
-DCMAKE_INSTALL_PREFIX="~/.local" will install in
~/.local/{bin,lib,include}, while -DCMAKE_INSTALL_PREFIX="~/.local/usr/"
will install in ~/.local/usr/{bin,lib,include}.