chore: mark the throw__ C helper noreturn; use assrt_not_null in place of actual asserts
This commit replaces some old actual asserts checking that pointer is
not null with calls to the assert_not_null helper. The helper exits the
program of the pointer is null. Exits, since C does not have exceptions
that can be thrown. Do note it is not expected the poiter is ever null
in these instancces; if it is, it is either a system problem/memory
corruption/invalid invocation.