refactor: add sched and filters headers and source files.
Begin implementation of some basic scheduler and filter interfaces,
inspired by Linux's queueing discpline system.

Replace the abstract base classes and dynamic casts with templates.
The abstraction for the QueueItem was leaky since the scheduler would
always need to downcast. Abstraction is therefore unsuitable and the
downcast a code smell, not to mention the overhead incurred by
dynamic_cast. Instead, use the appropriate concrete type required by a
scheduler.

The schedulers, filters etc are meant, in more complex configurations,
to be organized in a hierarchy, similar to `tc` qdisc setups.