chore(sched.hxx): remove improper use of forwarding references
by using a forwarding reference in the template functions here we can
easily end up with m_f as a reference to a local variable (imagine
we get called in a loop with some stack variable for the task) that
goes out of scope, causing a crash. Instead, the task should be taken
by value rather than reference, and then moved internally as appropriate.