perf: optmize sha256 implementation (~10% speedup)
unroll a few loops, and skip an extra mempcy when possible to process in
whole blocks. Also slight refactoring.

Note this implementation is ~4-5 slower than openssl's implementation
or the sha256sum command, likely especially because they make use
of the SHA-NI instructions on the machine this has been tested on,
thus benefitting from hardware offload.